==> notebooks/02_feature_importance.ipynb <==
# 1. Info

This notebooks contains the feature importance analysis.

Before runing this notebook you should have run the "__data_preparation.ipynb__" notebook.

# 2. Feature importance analysis.

## 2.1. Import Libraries

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import mutual_info_score
from IPython.display import display

## 2.2. Read the data

data = pd.read_csv('../data/enriched_data/premier_league.csv')

data.columns = data.columns.str.lower().str.replace(' ', '_')

## 2.3. Setting up the validation framework

categorical_variables = ['team','opponent','season','home']
numerical_variables = ['ftg_scored_total','ftg_received_total','htg_scored_total','htg_received_total','shots_total','shots_received_total','shots_target_total','shots_target_received_total','fouls_commited_total','fouls_received_total','corners_total','corners_against_total','yellow_cards_total','yellow_cards_opponent_total','red_cards_total','red_cards_opponent_total','points','goal_difference','position','win_rate','mooving_win_rate','mooving_goals_scored','mooving_goals_received']
# target_variable = ['ftg_scored','ftg_received','ft_result','win']

df = data[categorical_variables + numerical_variables + ['win']].copy()

df_full_train, df_test = train_test_split(df, test_size=0.2, random_state=1)
df_train, df_val = train_test_split(df_full_train, test_size=0.25, random_state=1)

print(len(df_train))
print(len(df_test))
print(len(df_val))

df_train = df_train.reset_index(drop=True)
df_val = df_val.reset_index(drop=True)
df_test = df_test.reset_index(drop=True)

y_train = df_train.win.values
y_val = df_val.win.values
y_test = df_test.win.values

del df_train['win']
del df_val['win']
del df_test['win']

## 2.4 difference and risk ratio

global_win = df_full_train.win.mean()

for cat in categorical_variables:
    print(cat)
    df_group = df_full_train.groupby(cat).win.agg(['mean', 'count'])
    df_group['diff'] = df_group['mean'] - global_win
    df_group['risk_ratio'] = df_group['mean'] / global_win
    display(df_group)
    print()
    print()

## 2.5. Mutual information (Categorical variables)

def mutual_info_churn_score(series):
    return mutual_info_score(series, df_full_train.win)

mi = df_full_train[categorical_variables].apply(mutual_info_churn_score)
mi.sort_values(ascending=False)

## 2.6 Correlation (Numerical variables)

numerical_corr = df_full_train[numerical_variables].corrwith(df_full_train.win).abs().reset_index()
numerical_corr.columns = ['features','corr']
numerical_corr.sort_values(by=['corr'], ascending=False)

numerical_corr.query(f"corr > {numerical_corr['corr'].mean()}")['features'].values

End of the notebook
