==> notebooks/07_predict.ipynb <==
# 1. Info

Notebook to make predictions using the ¨__xgboost_model.bin__" file.

Before running this notebook run the notebook "__train_model.ipynb__".

# 2. Make predictions

import pickle
import pandas as pd
import xgboost as xgb

## 2.1 Read model

model_file = '../xgboost_model.pkl'
dv_file = '../dv.pkl'

# read the model
with open(model_file, 'rb') as f_in:
    model = pickle.load(f_in)

# read the dv
with open(dv_file, 'rb') as f_in:
    dv = pickle.load(f_in)

## 2.2 Read data to predict

data = {
    "ftg_scored_total": 12,
    "htg_scored_total": 5,
    "points": 13,
    "goal_difference": 8,
    "position": 1,
    "win_rate": 0.8,
    "mooving_win_rate": 0.8,
    "mooving_goals_scored": 6.4
}

## 2.3 Vectorize the data

X = dv.fit_transform(data)

features = dv.feature_names_

dX = xgb.DMatrix(X, feature_names=features)

## 2.4 Make predictions

y_pred = model.predict(dX)[0]

bool(y_pred >= 0.5)

y_pred

End of the notebook
