==> notebooks/08_test_service.ipynb <==
# 1 Info

Notebook to test the services. Local and cloud.

# 2. Import Requests

import requests

# 3. Set jsons to test the service.

team_data = {
    "ftg_scored_total": 12,
    "htg_scored_total": 5,
    "points": 13,
    "goal_difference": 8,
    "position": 1,
    "win_rate": 0.8,
    "mooving_win_rate": 0.8,
    "mooving_goals_scored": 6.4
 }

team_data_2 = {
    "ftg_scored_total": 12,
    "htg_scored_total": 5,
    "points": 13,
    "goal_difference": 20,
    "position": 1,
    "win_rate": 0.8,
    "mooving_win_rate": 0.8,
    "mooving_goals_scored": 9.4
}

# 4. Test local service

This is sample code to test the service locally. There are two ways to activate the service locally:

[Using gunicorn inside the virutal environment](https://github.com/sergioasb8/midterm_project/tree/main#virtual-enviroment)

[Using docker](https://github.com/sergioasb8/midterm_project/tree/main#virtual-enviroment)
    
Follow the __README.md__ to activate the service on your computer with the option of your preference. 

local_url = 'http://0.0.0.0:9696/football_result'

requests.post(local_url, json=team_data).json()

requests.post(local_url, json=team_data_2).json()

# 5. Test cloud service

This service is live in the cloud but since is using a free plan from render, the first call could take longer than expected. Please be patient.

cloud_url = 'https://footballpredictor.onrender.com/football_result'

requests.post(cloud_url, json=team_data).json()

requests.post(cloud_url, json=team_data_2).json()
